Refactor Dataset component
