+++ crates/runtime/src/component/dataset/mod.rs
@@ -31,10 +31,13 @@ use snafu::prelude::*;
 use spicepod::{
     component::{dataset as spicepod_dataset, embeddings::ColumnEmbeddingConfig},
     metric::Metrics,
-    param::Params,
     semantic::Column,
 };
 use std::{collections::HashMap, fmt::Display, str::FromStr, sync::Arc, time::Duration};
+
+pub mod acceleration;
+pub mod builder;
+pub mod replication;
 
 #[derive(Debug, Snafu)]
 pub enum Error {
@@ -275,184 +278,6 @@ impl PartialEq for Dataset {
             && self.schema == other.schema
             && self.columns == other.columns
             && self.metrics == other.metrics
-    }
-}
-
-pub struct DatasetBuilder {
-    pub from: String,
-    pub name: TableReference,
-    pub mode: Mode,
-    pub params: HashMap<String, String>,
-    pub metadata: HashMap<String, String>,
-    pub columns: Vec<Column>,
-    pub has_metadata_table: bool,
-    pub replication: Option<replication::Replication>,
-    pub time_column: Option<String>,
-    pub time_format: Option<TimeFormat>,
-    pub time_partition_column: Option<String>,
-    pub time_partition_format: Option<TimeFormat>,
-    pub acceleration: Option<acceleration::Acceleration>,
-    pub embeddings: Vec<ColumnEmbeddingConfig>,
-    pub app: Option<Arc<App>>,
-    pub unsupported_type_action: Option<UnsupportedTypeAction>,
-    pub ready_state: ReadyState,
-    pub metrics: Metrics,
-    pub runtime: Option<Arc<Runtime>>,
-}
-
-impl TryFrom<spicepod_dataset::Dataset> for DatasetBuilder {
-    type Error = crate::Error;
-
-    fn try_from(dataset: spicepod_dataset::Dataset) -> std::result::Result<Self, Self::Error> {
-        #[allow(deprecated)]
-        let ready_state = match dataset.acceleration.as_ref().map(|a| a.ready_state) {
-            Some(Some(ready_state)) => {
-                tracing::warn!(
-                    "{}: `dataset.acceleration.ready_state` is deprecated, use `dataset.ready_state` instead.",
-                    dataset.name
-                );
-                ReadyState::from(ready_state)
-            }
-            _ => ReadyState::from(dataset.ready_state),
-        };
-
-        let acceleration = dataset
-            .acceleration
-            .map(acceleration::Acceleration::try_from)
-            .transpose()?;
-
-        validate_identifier(&dataset.name).context(crate::ComponentSnafu)?;
-
-        let table_reference = Dataset::parse_table_reference(&dataset.name)?;
-
-        Ok(DatasetBuilder {
-            from: dataset.from,
-            name: table_reference,
-            mode: Mode::from(dataset.mode),
-            params: dataset
-                .params
-                .as_ref()
-                .map(Params::as_string_map)
-                .unwrap_or_default(),
-            metadata: dataset
-                .metadata
-                .iter()
-                .map(|(k, v)| (k.clone(), v.to_string()))
-                .collect(),
-            columns: dataset.columns,
-            has_metadata_table: dataset
-                .has_metadata_table
-                .unwrap_or(DatasetBuilder::have_metadata_table_by_default()),
-            replication: dataset.replication.map(replication::Replication::from),
-            time_column: dataset.time_column,
-            time_format: dataset.time_format.map(TimeFormat::from),
-            time_partition_column: dataset.time_partition_column,
-            time_partition_format: dataset.time_partition_format.map(TimeFormat::from),
-            embeddings: dataset.embeddings,
-            acceleration,
-            app: None,
-            unsupported_type_action: dataset
-                .unsupported_type_action
-                .map(UnsupportedTypeAction::from),
-            ready_state,
-            metrics: dataset.metrics.unwrap_or_default(),
-            runtime: None,
-        })
-    }
-}
-
-impl DatasetBuilder {
-    pub fn try_new(from: String, name: &str) -> std::result::Result<Self, crate::Error> {
-        Ok(DatasetBuilder {
-            from,
-            name: Self::parse_table_reference(name)?,
-            mode: Mode::default(),
-            params: HashMap::default(),
-            metadata: HashMap::default(),
-            columns: Vec::default(),
-            has_metadata_table: Self::have_metadata_table_by_default(),
-            replication: None,
-            time_column: None,
-            time_format: None,
-            time_partition_column: None,
-            time_partition_format: None,
-            acceleration: None,
-            embeddings: Vec::default(),
-            app: None,
-            unsupported_type_action: None,
-            ready_state: ReadyState::default(),
-            metrics: Metrics::default(),
-            runtime: None,
-        })
-    }
-
-    pub(crate) fn parse_table_reference(
-        name: &str,
-    ) -> std::result::Result<TableReference, crate::Error> {
-        match TableReference::parse_str(name) {
-            table_ref @ (TableReference::Bare { .. } | TableReference::Partial { .. }) => {
-                Ok(table_ref)
-            }
-            TableReference::Full { catalog, .. } => crate::DatasetNameIncludesCatalogSnafu {
-                catalog,
-                name: name.to_string(),
-            }
-            .fail(),
-        }
-    }
-
-    #[must_use]
-    /// Returns whether the dataset should enable metadata by default.
-    fn have_metadata_table_by_default() -> bool {
-        false
-    }
-
-    #[must_use]
-    pub fn with_app(mut self, app: Arc<App>) -> Self {
-        self.app = Some(app);
-        self
-    }
-
-    #[must_use]
-    pub fn with_runtime(mut self, runtime: Arc<Runtime>) -> Self {
-        self.runtime = Some(runtime);
-        self
-    }
-
-    pub fn build(self) -> Result<Dataset> {
-        let app = self.app.ok_or(Error::UnableToBuildDataset {
-            dataset: self.name.to_string(),
-            missing_component: "app".to_string(),
-        })?;
-        let runtime = self.runtime.ok_or(Error::UnableToBuildDataset {
-            dataset: self.name.to_string(),
-            missing_component: "runtime".to_string(),
-        })?;
-
-        let dataset = Dataset {
-            from: self.from,
-            name: self.name,
-            mode: self.mode,
-            params: self.params,
-            metadata: self.metadata,
-            columns: self.columns,
-            has_metadata_table: self.has_metadata_table,
-            replication: self.replication,
-            time_column: self.time_column,
-            time_format: self.time_format,
-            time_partition_column: self.time_partition_column,
-            time_partition_format: self.time_partition_format,
-            acceleration: self.acceleration,
-            embeddings: self.embeddings,
-            app,
-            schema: None,
-            unsupported_type_action: self.unsupported_type_action,
-            ready_state: self.ready_state,
-            metrics: self.metrics,
-            runtime,
-        };
-
-        Ok(dataset)
     }
 }
 
@@ -768,25 +593,6 @@ impl AccelerationSource for Dataset {
     }
 }
 
-pub mod acceleration;
-
-pub mod replication {
-    use spicepod::component::dataset::replication as spicepod_replication;
-
-    #[derive(Debug, Clone, PartialEq, Default)]
-    pub struct Replication {
-        pub enabled: bool,
-    }
-
-    impl From<spicepod_replication::Replication> for Replication {
-        fn from(replication: spicepod_replication::Replication) -> Self {
-            Replication {
-                enabled: replication.enabled,
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
@@ -794,6 +600,7 @@ mod tests {
     use datafusion_table_providers::util::column_reference::ColumnReference;
 
     use super::acceleration::{Acceleration, IndexType};
+    use super::builder::DatasetBuilder;
     use super::*;
     use app::AppBuilder;
