+++ crates/runtime/src/component/dataset/replication.rs
@@ -1,0 +1,30 @@
+/*
+Copyright 2025 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+use spicepod::component::dataset::replication as spicepod_replication;
+
+#[derive(Debug, Clone, PartialEq, Default)]
+pub struct Replication {
+    pub enabled: bool,
+}
+
+impl From<spicepod_replication::Replication> for Replication {
+    fn from(replication: spicepod_replication::Replication) -> Self {
+        Replication {
+            enabled: replication.enabled,
+        }
+    }
+}
