+++ crates/runtime/src/internal_table.rs
@@ -24,8 +24,7 @@ use tokio::sync::RwLock;
 
 use crate::Runtime;
 use crate::accelerated_table::{AcceleratedTableBuilderError, Retention};
-use crate::component::dataset::acceleration::Acceleration;
-use crate::component::dataset::{DatasetBuilder, Mode};
+use crate::component::dataset::{Mode, acceleration::Acceleration, builder::DatasetBuilder};
 use crate::federated_table::FederatedTable;
 use crate::secrets::Secrets;
 use crate::status;
